fixed int -> double

+++ Parcel.cs
@@ -6,8 +6,8 @@ namespace WebAPIParcelTracking
     {
         [BsonId]
         public string? Id { get; set; }
-        public int? Lat { get; set; }
-        public int? Long { get; set; }
+        public double? Lat { get; set; }
+        public double? Long { get; set; }
         public double? Weight { get; set; }
         public string? Desc { get; set; }
     }
